fix: Deploy and Publish integration tests - patched boto client
Why is this change necessary?

* Due to eventual consistency on S3, there are sometimes transient
failures when the same object is tried to be retrieved after just being
uploaded.

How does it address the issue?

* The fix allows for a blocking call to s3 `put_object`, which makes
get_object calls to ensure that a given object actually exists in S3.
* `get_object` calls

What side effects does this change have?

* The boto client in the integration tests alone have this consistency
check, because thats where the patch happens.